+++ parsidate/holidays/__init__.py
@@ -1,4 +1,4 @@
-"""Iranian official holiday data and HolidaySet container.
+"""Iranian official holiday data and calendar containers.
 
 Copyright (C) 2024 Ali Sadeghi Aghili
 Licensed under the Apache License, Version 2.0
@@ -17,11 +17,13 @@ from parsidate.holidays.ir import (
     is_holiday,
     iran_holidays,
 )
+from parsidate.holidays.calendar import HolidayCalendar
 
 __all__ = [
     "DATA_VERSION",
     "HolidayDataMissing",
     "HolidaySet",
+    "HolidayCalendar",
     "iran_fixed_solar_holidays",
     "iran_holidays",
     "holidays_in_year",